requirements.txt
anndata==0.11.1
annoy==1.17.3
fbpca==1.0
graphtools==1.5.3
hnswlib==0.8.0
matplotlib==3.7.5
numpy==1.22.3
pandas==2.0.3
pytz==2024.2
PyYAML==6.0.2
scanpy==1.10.4
scikit_learn==1.3.2
scipy==1.14.1
seaborn==0.13.2
torch==2.3.0
tqdm==4.66.5
joblib==1.4.2
networkx==3.1